polish: Final project polish - fix warnings and update README (US-309)
Changes:
- Fixed compiler warning in anomaly_detector.cpp (sign comparison)
  * Added static_cast<size_t> for proper type comparison
  * Eliminates all compiler warnings
- Updated main README.md with professional structure
  * Added badges (License, C++17, Java)
  * Added comprehensive documentation links to docs/final
  * Added performance metrics and tech stack tables
  * Added quick start code examples
  * Clear navigation to all guides and API docs
- All tests passing (C++ and Java)
- Project ready for presentation

+++ stream-processor/src/anomaly_detector.cpp
@@ -144,7 +144,7 @@ std::optional<AnomalyResult> AnomalyDetector::analyze(const Event& event) {
         baseline.update(event);
 
         // Check if we have enough events to establish baseline
-        if (baseline.total_events >= min_events_for_baseline_) {
+        if (static_cast<size_t>(baseline.total_events) >= min_events_for_baseline_) {
             baseline.is_baseline_ready = true;
         }
